Fixed level identification problem with proc models
At first I thought it was because of the "BOX" variable that I did not take it into consideration, even so the problems continued then ...

The try block did not take into consideration the if of the models involved. "Gb800se", "gb800solo", "gb800ue"

Before in GBtrio4k try it tried to open led0 and led1 when this model does not correspond with the proc characteristics. As a consequence a continuous log "[LED-GIGA] set LED Pattern failed!"

Now if you take into consideration the models involved.